+++ tools/src/h5stat/h5stat.c
@@ -1603,7 +1603,7 @@ print_storage_summary(const iter_t *iter)
 {
     hsize_t total_meta = 0;
     hsize_t unaccount  = 0;
-    double  percent    = 0.0f;
+    double  percent    = 0.0;
 
     HDfprintf(stdout, "File space management strategy: %s\n", FS_STRATEGY_NAME[iter->fs_strategy]);
     HDfprintf(stdout, "File space page size: %" PRIuHSIZE " bytes\n", iter->fsp_size);
@@ -1618,7 +1618,7 @@ print_storage_summary(const iter_t *iter)
     HDfprintf(stdout, "  File metadata: %" PRIuHSIZE " bytes\n", total_meta);
     HDfprintf(stdout, "  Raw data: %" PRIuHSIZE " bytes\n", iter->dset_storage_size);
 
-    percent = ((double)iter->free_space / (double)iter->filesize) * (double)100.0f;
+    percent = ((double)iter->free_space / (double)iter->filesize) * 100.0;
     HDfprintf(stdout, "  Amount/Percent of tracked free space: %" PRIuHSIZE " bytes/%3.1f%%\n",
               iter->free_space, percent);
